+++ relay-server/src/service.rs
@@ -16,6 +16,7 @@ use tokio::runtime::Runtime;
 use crate::services::cogs::{CogsService, CogsServiceRecorder};
 use crate::services::global_config::{GlobalConfigManager, GlobalConfigService};
 use crate::services::health_check::{HealthCheck, HealthCheckService};
+use crate::services::metric_stats::{MetricStatsService, TrackMetric};
 use crate::services::outcome::{OutcomeProducer, OutcomeProducerService, TrackOutcome};
 use crate::services::outcome_aggregator::OutcomeAggregator;
 use crate::services::processor::{EnvelopeProcessor, EnvelopeProcessorService};
@@ -50,6 +51,7 @@ pub struct Registry {
     pub health_check: Addr<HealthCheck>,
     pub outcome_producer: Addr<OutcomeProducer>,
     pub outcome_aggregator: Addr<TrackOutcome>,
+    pub metric_stats: Addr<TrackMetric>,
     pub processor: Addr<EnvelopeProcessor>,
     pub test_store: Addr<TestStore>,
     pub relay_cache: Addr<RelayCache>,
@@ -136,6 +138,8 @@ impl ServiceState {
         )
         .start_in(&runtimes.aggregator);
 
+        let metric_stats = MetricStatsService::new(config.clone(), aggregator.clone()).start();
+
         #[cfg(feature = "processing")]
         let store = match runtimes.store {
             Some(ref rt) => Some(
@@ -143,6 +147,7 @@ impl ServiceState {
                     config.clone(),
                     global_config_handle.clone(),
                     outcome_aggregator.clone(),
+                    metric_stats.clone(),
                 )?
                 .start_in(rt),
             ),
@@ -214,6 +219,7 @@ impl ServiceState {
             health_check,
             outcome_producer,
             outcome_aggregator,
+            metric_stats,
             test_store,
             relay_cache,
             global_config,