+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Track metric bucket metadata for Relay internal usage. ([#3254](https://github.com/getsentry/relay/pull/3254))
 - Enforce rate limits for standalone spans. ([#3238](https://github.com/getsentry/relay/pull/3238))
 - Extract `span.status_code` tag for HTTP spans. ([#3245](https://github.com/getsentry/relay/pull/3245))
+- Implement volume metric stats. ([#3245](https://github.com/getsentry/relay/pull/3245))
 
 **Bug Fixes**: